rename modules directories
