+++ nlisim/modules/hemoglobin.py
@@ -6,7 +6,7 @@ import numpy as np
 
 from nlisim.coordinates import Voxel
 from nlisim.grid import RectangularGrid
-from nlisim.henrique_modules.molecules import MoleculeModel, MoleculesState
+from nlisim.modules.molecules import MoleculeModel, MoleculesState
 from nlisim.module import ModuleState
 from nlisim.state import State
 from nlisim.util import turnover_rate
@@ -42,7 +42,7 @@ class Hemoglobin(MoleculeModel):
 
     def advance(self, state: State, previous_time: float) -> State:
         """Advance the state by a single time step."""
-        from nlisim.henrique_modules.afumigatus import (
+        from nlisim.modules.afumigatus import (
             AfumigatusCellData,
             AfumigatusState,
             AfumigatusCellStatus,
