+++ nlisim/modules/il6.py
@@ -6,7 +6,7 @@ import numpy as np
 
 from nlisim.coordinates import Voxel
 from nlisim.grid import RectangularGrid
-from nlisim.henrique_modules.molecules import MoleculeModel, MoleculesState
+from nlisim.modules.molecules import MoleculeModel, MoleculesState
 from nlisim.module import ModuleState
 from nlisim.state import State
 from nlisim.util import turnover_rate
@@ -57,9 +57,9 @@ class IL6(MoleculeModel):
 
     def advance(self, state: State, previous_time: float) -> State:
         """Advance the state by a single time step."""
-        from nlisim.henrique_modules.macrophage import MacrophageState
-        from nlisim.henrique_modules.neutrophil import NeutrophilState
-        from nlisim.henrique_modules.phagocyte import PhagocyteStatus
+        from nlisim.modules.macrophage import MacrophageState
+        from nlisim.modules.neutrophil import NeutrophilState
+        from nlisim.modules.phagocyte import PhagocyteStatus
 
         il6: IL6State = state.il6
         molecules: MoleculesState = state.molecules
