+++ nlisim/modules/lactoferrin.py
@@ -5,7 +5,7 @@ import numpy as np
 
 from nlisim.coordinates import Voxel
 from nlisim.grid import RectangularGrid
-from nlisim.henrique_modules.molecules import MoleculeModel, MoleculesState
+from nlisim.modules.molecules import MoleculeModel, MoleculesState
 from nlisim.module import ModuleState
 from nlisim.state import State
 from nlisim.util import iron_tf_reaction, michaelian_kinetics, turnover_rate
@@ -62,11 +62,11 @@ class Lactoferrin(MoleculeModel):
 
     def advance(self, state: State, previous_time: float) -> State:
         """Advance the state by a single time step."""
-        from nlisim.henrique_modules.neutrophil import NeutrophilCellData, NeutrophilState
-        from nlisim.henrique_modules.phagocyte import PhagocyteState, PhagocyteStatus
-        from nlisim.henrique_modules.macrophage import MacrophageCellData, MacrophageState
-        from nlisim.henrique_modules.transferrin import TransferrinState
-        from nlisim.henrique_modules.iron import IronState
+        from nlisim.modules.neutrophil import NeutrophilCellData, NeutrophilState
+        from nlisim.modules.phagocyte import PhagocyteState, PhagocyteStatus
+        from nlisim.modules.macrophage import MacrophageCellData, MacrophageState
+        from nlisim.modules.transferrin import TransferrinState
+        from nlisim.modules.iron import IronState
 
         lactoferrin: LactoferrinState = state.lactoferrin
         transferrin: TransferrinState = state.transferrin
