+++ nlisim/modules/mcp1.py
@@ -7,17 +7,17 @@ import numpy as np
 from nlisim.coordinates import Voxel
 from nlisim.grid import RectangularGrid
 from nlisim.module import ModuleState
-from nlisim.henrique_modules.molecules import MoleculeModel, MoleculesState
+from nlisim.modules.molecules import MoleculeModel, MoleculesState
 from nlisim.state import State
 from nlisim.util import turnover_rate
 
 
-def molecule_grid_factory(self: 'MIP1BState') -> np.ndarray:
+def molecule_grid_factory(self: 'MCP1State') -> np.ndarray:
     return np.zeros(shape=self.global_state.grid.shape, dtype=float)
 
 
 @attr.s(kw_only=True, repr=False)
-class MIP1BState(ModuleState):
+class MCP1State(ModuleState):
     grid: np.ndarray = attr.ib(default=attr.Factory(molecule_grid_factory, takes_self=True))
     half_life: float
     half_life_multiplier: float
@@ -28,39 +28,35 @@ class MIP1BState(ModuleState):
     k_d: float
 
 
-class MIP1B(MoleculeModel):
-    """MIP1B"""
+class MCP1(MoleculeModel):
+    """MCP1"""
 
-    name = 'mip1b'
-    StateClass = MIP1BState
+    name = 'mcp1'
+    StateClass = MCP1State
 
     def initialize(self, state: State) -> State:
-        mip1b: MIP1BState = state.mip1b
+        mcp1: MCP1State = state.mcp1
 
         # config file values
-        mip1b.half_life = self.config.getfloat('half_life')
-        mip1b.macrophage_secretion_rate = self.config.getfloat('macrophage_secretion_rate')
-        mip1b.pneumocyte_secretion_rate = self.config.getfloat('pneumocyte_secretion_rate')
-        mip1b.k_d = self.config.getfloat('k_d')
+        mcp1.half_life = self.config.getfloat('half_life')
+        mcp1.macrophage_secretion_rate = self.config.getfloat('macrophage_secretion_rate')
+        mcp1.pneumocyte_secretion_rate = self.config.getfloat('pneumocyte_secretion_rate')
+        mcp1.k_d = self.config.getfloat('k_d')
 
         # computed values
-        mip1b.half_life_multiplier = 1 + math.log(0.5) / (mip1b.half_life / self.time_step)
+        mcp1.half_life_multiplier = 1 + math.log(0.5) / (mcp1.half_life / self.time_step)
         # time unit conversions
-        mip1b.macrophage_secretion_rate_unit_t = (
-            mip1b.macrophage_secretion_rate * 60 * self.time_step
-        )
-        mip1b.pneumocyte_secretion_rate_unit_t = (
-            mip1b.pneumocyte_secretion_rate * 60 * self.time_step
-        )
+        mcp1.macrophage_secretion_rate_unit_t = mcp1.macrophage_secretion_rate * 60 * self.time_step
+        mcp1.pneumocyte_secretion_rate_unit_t = mcp1.pneumocyte_secretion_rate * 60 * self.time_step
 
         return state
 
     def advance(self, state: State, previous_time: float) -> State:
         """Advance the state by a single time step."""
-        from nlisim.henrique_modules.macrophage import MacrophageCellData, MacrophageState
-        from nlisim.henrique_modules.pneumocyte import PneumocyteCellData, PneumocyteState
+        from nlisim.modules.macrophage import MacrophageCellData, MacrophageState
+        from nlisim.modules.pneumocyte import PneumocyteCellData, PneumocyteState
 
-        mip1b: MIP1BState = state.mip1b
+        mcp1: MCP1State = state.mcp1
         molecules: MoleculesState = state.molecules
         pneumocyte: PneumocyteState = state.pneumocyte
         macrophage: MacrophageState = state.macrophage
@@ -72,7 +68,7 @@ class MIP1B(MoleculeModel):
 
             if pneumocyte_cell['tnfa']:
                 pneumocyte_cell_voxel: Voxel = grid.get_voxel(pneumocyte_cell['point'])
-                mip1b.grid[tuple(pneumocyte_cell_voxel)] += mip1b.pneumocyte_secretion_rate_unit_t
+                mcp1.grid[tuple(pneumocyte_cell_voxel)] += mcp1.pneumocyte_secretion_rate_unit_t
 
         # interact with macrophages
         for macrophage_cell_index in macrophage.cells.alive():
@@ -80,30 +76,30 @@ class MIP1B(MoleculeModel):
 
             if macrophage_cell['tnfa']:
                 macrophage_cell_voxel: Voxel = grid.get_voxel(macrophage_cell['point'])
-                mip1b.grid[tuple(macrophage_cell_voxel)] += mip1b.macrophage_secretion_rate_unit_t
+                mcp1.grid[tuple(macrophage_cell_voxel)] += mcp1.macrophage_secretion_rate_unit_t
 
-        # Degrade MIP1B
-        mip1b.grid *= mip1b.half_life_multiplier
-        mip1b.grid *= turnover_rate(
+        # Degrade MCP1
+        mcp1.grid *= mcp1.half_life_multiplier
+        mcp1.grid *= turnover_rate(
             x_mol=np.array(1.0, dtype=np.float64),
             x_system_mol=0.0,
             base_turnover_rate=molecules.turnover_rate,
             rel_cyt_bind_unit_t=molecules.rel_cyt_bind_unit_t,
         )
 
-        # Diffusion of MIP1b
-        self.diffuse(mip1b.grid, state)
+        # Diffusion of MCP1
+        self.diffuse(mcp1.grid, state)
 
         return state
 
     def summary_stats(self, state: State) -> Dict[str, Any]:
-        mip1b: MIP1BState = state.mip1b
+        mcp1: MCP1State = state.mcp1
         voxel_volume = state.voxel_volume
 
         return {
-            'concentration': float(np.mean(mip1b.grid) / voxel_volume),
+            'concentration': float(np.mean(mcp1.grid) / voxel_volume),
         }
 
     def visualization_data(self, state: State) -> Tuple[str, Any]:
-        mip1b: MIP1BState = state.mip1b
-        return 'molecule', mip1b.grid
+        mcp1: MCP1State = state.mcp1
+        return 'molecule', mcp1.grid
