+++ nlisim/modules/neutrophil.py
@@ -1,299 +1,254 @@
-from enum import IntEnum
-import itertools
-from random import choice, shuffle
+import math
 from typing import Any, Dict, Tuple
 
 import attr
+from attr import attrib, attrs
 import numpy as np
 
 from nlisim.cell import CellData, CellList
 from nlisim.coordinates import Point, Voxel
 from nlisim.grid import RectangularGrid
-from nlisim.module import ModuleModel, ModuleState
-from nlisim.modules.fungus import FungusCellData, FungusCellList
-from nlisim.util import TissueType
+from nlisim.modules.mip2 import MIP2State
+from nlisim.modules.phagocyte import (
+    internalize_aspergillus,
+    PhagocyteCellData,
+    PhagocyteModel,
+    PhagocyteModuleState,
+    PhagocyteState,
+    PhagocyteStatus,
+)
 from nlisim.random import rg
 from nlisim.state import State
+from nlisim.util import activation_function, TissueType
+
+MAX_CONIDIA = (
+    50  # note: this the max that we can set the max to. i.e. not an actual model parameter
+)
 
 
-class NeutrophilCellData(CellData):
-    class Status(IntEnum):
-        NONGRANULATING = 0
-        GRANULATING = 1
-
-    NEUTROPHIL_FIELDS = [('status', 'u1'), ('iteration', 'i4'), ('granule_count', 'i4')]
+class NeutrophilCellData(PhagocyteCellData):
+    NEUTROPHIL_FIELDS = [
+        ('status', np.uint8),
+        ('state', np.uint8),
+        ('iron_pool', np.float64),
+        ('max_move_step', np.float64),  # TODO: double check, might be int
+        ('tnfa', bool),
+        ('engaged', bool),
+        ('status_iteration', np.uint),
+    ]
 
     dtype = np.dtype(CellData.FIELDS + NEUTROPHIL_FIELDS, align=True)  # type: ignore
 
     @classmethod
     def create_cell_tuple(
         cls,
-        *,
-        status=Status.NONGRANULATING,
-        granule_count=0,
         **kwargs,
     ) -> np.record:
-        iteration = 0
-        return CellData.create_cell_tuple(**kwargs) + (
-            status,
-            iteration,
-            granule_count,
+        initializer = {
+            'status': kwargs.get('status', PhagocyteStatus.RESTING),
+            'state': kwargs.get('state', PhagocyteState.FREE),
+            'iron_pool': kwargs.get('iron_pool', 0.0),
+            'max_move_step': kwargs.get('max_move_step', 1.0),  # TODO: reasonable default?
+            'tnfa': kwargs.get('tnfa', False),
+            'engaged': kwargs.get('engaged', False),
+            'status_iteration': kwargs.get('status_iteration', 0),
+        }
+
+        # ensure that these come in the correct order
+        return CellData.create_cell_tuple(**kwargs) + tuple(
+            [initializer[key] for key, *_ in NeutrophilCellData.NEUTROPHIL_FIELDS]
         )
 
 
-@attr.s(kw_only=True, frozen=True, repr=False)
+@attrs(kw_only=True, frozen=True, repr=False)
 class NeutrophilCellList(CellList):
     CellDataClass = NeutrophilCellData
 
-    def recruit_new(self, rec_rate_ph, rec_r, granule_count, neutropenic, time, grid, tissue, cyto):
-        num_reps = 0
-        if not neutropenic:
-            num_reps = rec_rate_ph  # number of neutrophils recruited per time step
-        elif neutropenic and 48 <= time <= 96:
-            # TODO: relate 3 to Algorithm S3.14 and rec_rate_ph or introduce neutropenic parameter
-            #  In S3.14: num_reps = 6 and int( (time-48)/ 8), both are 1/3 values here
-            num_reps = int((time - 48) / 8) * 3
 
-        if num_reps <= 0:
-            return
-
-        cyto_index = np.argwhere(np.logical_and(tissue == TissueType.BLOOD.value, cyto >= rec_r))
-        if len(cyto_index) <= 0:
-            # nowhere to place cells
-            return
-
-        for _ in range(num_reps):
-            ii = rg.integers(cyto_index.shape[0])
-            point = Point(
-                x=grid.x[cyto_index[ii, 2]],
-                y=grid.y[cyto_index[ii, 1]],
-                z=grid.z[cyto_index[ii, 0]],
-            )
-
-            self.append(
-                NeutrophilCellData.create_cell(
-                    point=point,
-                    status=NeutrophilCellData.Status.NONGRANULATING,
-                    granule_count=granule_count,
-                )
-            )
-
-    def absorb_cytokines(self, n_absorb, cyto, grid):
-        for index in self.alive():
-            vox = grid.get_voxel(self[index]['point'])
-            x = vox.x
-            y = vox.y
-            z = vox.z
-            cyto[z, y, x] = (1 - n_absorb) * cyto[z, y, x]
-
-    def produce_cytokines(self, n_det, n_n, grid, fungus: FungusCellList, cyto):
-        for i in self.alive():
-            vox = grid.get_voxel(self[i]['point'])
-
-            hyphae_count = 0
-
-            # Moore neighborhood
-            neighborhood = tuple(itertools.product(tuple(range(-1 * n_det, n_det + 1)), repeat=3))
-
-            for dx, dy, dz in neighborhood:
-                zi = vox.z + dz
-                yj = vox.y + dy
-                xk = vox.x + dx
-                if grid.is_valid_voxel(Voxel(x=xk, y=yj, z=zi)):
-                    index_arr = fungus.get_cells_in_voxel(Voxel(x=xk, y=yj, z=zi))
-                    for index in index_arr:
-                        if fungus[index]['form'] == FungusCellData.Form.HYPHAE:
-                            hyphae_count += 1
-
-            cyto[vox.z, vox.y, vox.x] = cyto[vox.z, vox.y, vox.x] + (n_n * hyphae_count)
-
-    def move(self, rec_r, grid, cyto, tissue):
-        for cell_index in self.alive(
-            self.cell_data['status'] == NeutrophilCellData.Status.NONGRANULATING
-        ):
-            # TODO: Algorithm S3.17 says "if degranulating nearby hyphae, do not move" but do
-            #  we have the "nearby hyphae" part of this condition?
-            cell = self[cell_index]
-            cell_voxel = grid.get_voxel(cell['point'])
-
-            valid_voxel_offsets = []
-            above_threshold_voxel_offsets = []
-
-            # iterate over nearby voxels, recording the cytokine levels
-            for dx, dy, dz in itertools.product((-1, 0, 1), repeat=3):
-                zi = cell_voxel.z + dz
-                yj = cell_voxel.y + dy
-                xk = cell_voxel.x + dx
-                if grid.is_valid_voxel(Voxel(x=xk, y=yj, z=zi)):
-                    if tissue[zi, yj, xk] != TissueType.AIR.value:
-                        valid_voxel_offsets.append((dx, dy, dz))
-                        if cyto[zi, yj, xk] >= rec_r:
-                            above_threshold_voxel_offsets.append((cyto[zi, yj, xk], (dx, dy, dz)))
-
-            # pick a target for the move
-            if len(above_threshold_voxel_offsets) > 0:
-                # shuffle + sort (with _only_ 0-key, not lexicographic as tuples) ensures
-                # randomization when there are equal top cytokine levels
-                # note that numpy's shuffle will complain about ragged arrays
-                shuffle(above_threshold_voxel_offsets)
-                above_threshold_voxel_offsets = sorted(
-                    above_threshold_voxel_offsets, key=lambda x: x[0], reverse=True
-                )
-                _, target_voxel_offset = above_threshold_voxel_offsets[0]
-            elif len(valid_voxel_offsets) > 0:
-                target_voxel_offset = choice(valid_voxel_offsets)
-            else:
-                raise AssertionError(
-                    'This cell has no valid voxel to move to, including the one that it is in!'
-                )
-
-            # Some nonsense here, b/c jump is happening at the voxel level, not the point level
-            starting_cell_point = Point(x=cell['point'][2], y=cell['point'][1], z=cell['point'][0])
-            starting_cell_voxel = grid.get_voxel(starting_cell_point)
-            ending_cell_voxel = grid.get_voxel(
-                Point(
-                    x=grid.x[cell_voxel.x + target_voxel_offset[0]],
-                    y=grid.y[cell_voxel.y + target_voxel_offset[1]],
-                    z=grid.z[cell_voxel.z + target_voxel_offset[2]],
-                )
-            )
-            ending_cell_point = (
-                starting_cell_point
-                + grid.get_voxel_center(ending_cell_voxel)
-                - grid.get_voxel_center(starting_cell_voxel)
-            )
-
-            cell['point'] = ending_cell_point
-            self.update_voxel_index([cell_index])
-
-    def damage_hyphae(self, n_det, n_kill, time, health, grid, fungus: FungusCellList, iron):
-        for i in self.alive(self.cell_data['granule_count'] > 0):
-            cell = self[i]
-            vox = grid.get_voxel(cell['point'])
-
-            # Moore neighborhood, but order partially randomized. Closest to furthest order, but
-            # the order of any set of points of equal distance is random
-            neighborhood = list(itertools.product(tuple(range(-1 * n_det, n_det + 1)), repeat=3))
-            shuffle(neighborhood)
-            neighborhood = sorted(neighborhood, key=lambda v: v[0] ** 2 + v[1] ** 2 + v[2] ** 2)
-
-            for dx, dy, dz in neighborhood:
-                zi = vox.z + dz
-                yj = vox.y + dy
-                xk = vox.x + dx
-                if grid.is_valid_voxel(Voxel(x=xk, y=yj, z=zi)):
-                    index_arr = fungus.get_cells_in_voxel(Voxel(x=xk, y=yj, z=zi))
-                    if len(index_arr) > 0:
-                        iron[zi, yj, xk] = 0
-                    for index in index_arr:
-                        if (
-                            fungus[index]['form'] == FungusCellData.Form.HYPHAE
-                            and cell['granule_count'] > 0
-                        ):
-                            fungus[index]['health'] -= health * (time / n_kill)
-                            cell['granule_count'] -= 1
-                            cell['status'] = NeutrophilCellData.Status.GRANULATING
-                        elif cell['granule_count'] == 0:
-                            cell['status'] = NeutrophilCellData.Status.NONGRANULATING
-                            break
-
-    def update(self):
-        for i in self.alive(self.cell_data['granule_count'] == 0):
-            self[i]['status'] = NeutrophilCellData.Status.NONGRANULATING
-
-    def age(self):
-        self.cell_data['iteration'] += 1
-
-    def kill_by_age(self, age_limit):
-        for i in self.alive(self.cell_data['iteration'] > age_limit):
-            self[i]['dead'] = True
-
-
-def cell_list_factory(self: 'NeutrophilState'):
+def cell_list_factory(self: 'NeutrophilState') -> NeutrophilCellList:
     return NeutrophilCellList(grid=self.global_state.grid)
 
 
-@attr.s(kw_only=True)
-class NeutrophilState(ModuleState):
-    cells: NeutrophilCellList = attr.ib(default=attr.Factory(cell_list_factory, takes_self=True))
-    neutropenic: bool
-    rec_rate_ph: int
-    rec_r: float
-    n_absorb: float
-    n_n: float
-    n_det: int
-    granule_count: int
-    n_kill: float
-    time_n: float
-    age_limit: int
+@attrs(kw_only=True)
+class NeutrophilState(PhagocyteModuleState):
+    cells: NeutrophilCellList = attrib(default=attr.Factory(cell_list_factory, takes_self=True))
+    half_life: float
+    time_to_change_state: float
+    iter_to_change_state: int
+    pr_n_hyphae: float
+    pr_n_phagocyte: float
+    recruitment_rate: float
+    rec_bias: float
+    max_n: float  # TODO: 0.5?
+    n_frac: float
+    drift_bias: float
+    n_move_rate_act: float
+    n_move_rate_rest: float
 
 
-class Neutrophil(ModuleModel):
+class Neutrophil(PhagocyteModel):
     name = 'neutrophil'
-
     StateClass = NeutrophilState
 
     def initialize(self, state: State):
         neutrophil: NeutrophilState = state.neutrophil
-        grid: RectangularGrid = state.grid
+        voxel_volume = state.voxel_volume
+        time_step_size: float = self.time_step
 
-        neutrophil.neutropenic = self.config.getboolean('neutropenic')
-        neutrophil.rec_rate_ph = self.config.getint('rec_rate_ph')
-        neutrophil.rec_r = self.config.getfloat('rec_r')
-        neutrophil.n_absorb = self.config.getfloat('n_absorb')
-        neutrophil.n_n = self.config.getfloat('Nn')
-        neutrophil.n_det = self.config.getint('n_det')
-        neutrophil.granule_count = self.config.getint('granule_count')
-        neutrophil.n_kill = self.config.getfloat('n_kill')
-        neutrophil.time_n = self.config.getfloat('time_n')
-        neutrophil.age_limit = self.config.getint('age_limit')
+        neutrophil.init_num_neutrophils = self.config.getint('init_num_neutrophils')
 
-        neutrophil.cells = NeutrophilCellList(grid=grid)
+        neutrophil.time_to_change_state = self.config.getfloat('time_to_change_state')
+        neutrophil.max_conidia = self.config.getint('max_conidia')
+
+        neutrophil.recruitment_rate = self.config.getfloat('recruitment_rate')
+        neutrophil.rec_bias = self.config.getfloat('rec_bias')
+        neutrophil.max_n = self.config.getfloat('max_n')
+        neutrophil.n_frac = self.config.getfloat('n_frac')
+
+        neutrophil.drift_bias = self.config.getfloat('drift_bias')
+        neutrophil.n_move_rate_act = self.config.getfloat('n_move_rate_act')
+        neutrophil.n_move_rate_rest = self.config.getfloat('n_move_rate_rest')
+
+        # computed values
+        # TODO: not a real half life
+        neutrophil.half_life = -math.log(0.5) / (
+            self.config.getfloat('half_life') * (60 / time_step_size)
+        )
+
+        neutrophil.iter_to_change_state = int(neutrophil.time_to_change_state * 60 / time_step_size)
+
+        rel_n_hyphae_int_unit_t = time_step_size / 60  # per hour # TODO: not like this
+        neutrophil.pr_n_hyphae = 1 - math.exp(
+            -rel_n_hyphae_int_unit_t / (voxel_volume * self.config.getfloat('pr_n_hyphae_param'))
+        )  # TODO: -exp1m
+
+        rel_phagocyte_affinity_unit_t = time_step_size / 60  # TODO: not like this
+        neutrophil.pr_n_phagocyte = 1 - math.exp(
+            -rel_phagocyte_affinity_unit_t
+            / (voxel_volume * self.config.getfloat('pr_n_phag_param'))
+        )  # TODO: -exp1m
+
+        # place initial neutrophils
+        z_range, y_range, x_range = state.lung_tissue.shape
+        for _ in range(neutrophil.init_num_neutrophils):
+            z, y, x = rg.integers([z_range, y_range, x_range])
+            while state.lung_tissue[z, y, x] == TissueType.AIR:
+                z, y, x = rg.integers([z_range, y_range, x_range])
+
+            self.create_neutrophil(state=state, x=x, y=y, z=z)
 
         return state
 
     def advance(self, state: State, previous_time: float):
+        """Advance the state by a single time step."""
+        from nlisim.modules.afumigatus import (
+            AfumigatusCellData,
+            AfumigatusCellStatus,
+            AfumigatusState,
+        )
+        from nlisim.modules.iron import IronState
+        from nlisim.modules.macrophage import MacrophageCellData, MacrophageState
+
         neutrophil: NeutrophilState = state.neutrophil
-        n_cells = neutrophil.cells
-        fungus = state.fungus.cells
-        health = state.fungus.health
+        macrophage: MacrophageState = state.macrophage
+        afumigatus: AfumigatusState = state.afumigatus
+        iron: IronState = state.iron
+        grid: RectangularGrid = state.grid
+        voxel_volume: float = state.voxel_volume
+        space_volume: float = state.space_volume
 
-        tissue = state.geometry.lung_tissue
-        grid = state.grid
-        cyto = state.molecules.grid['n_cyto']
-        iron = state.molecules.grid['iron']
+        for neutrophil_cell_index in neutrophil.cells.alive():
+            neutrophil_cell = neutrophil.cells[neutrophil_cell_index]
+            neutrophil_cell_voxel: Voxel = grid.get_voxel(neutrophil_cell['point'])
 
-        # recruit new
-        n_cells.recruit_new(
-            neutrophil.rec_rate_ph,
-            neutrophil.rec_r,
-            neutrophil.granule_count,
-            neutrophil.neutropenic,
-            previous_time,
-            grid,
-            tissue,
-            cyto,
-        )
+            self.update_status(state, neutrophil_cell)
 
-        # absorb cytokines
-        n_cells.absorb_cytokines(neutrophil.n_absorb, cyto, grid)
+            neutrophil_cell['move_step'] = 0
+            # TODO: -1 below was 'None'. this looks like something which needs to be reworked
+            neutrophil_cell['max_move_step'] = -1
+            neutrophil_cell['engaged'] = False  # TODO: find out what 'engaged' means
 
-        # produce cytokines
-        n_cells.produce_cytokines(neutrophil.n_det, neutrophil.n_n, grid, fungus, cyto)
+            # ---------- interactions
 
-        # move
-        n_cells.move(neutrophil.rec_r, grid, cyto, tissue)
+            # dead and dying cells release iron
+            # TODO: can move this to a numpy operation if it ends up more performant
+            if neutrophil_cell['status'] in {
+                PhagocyteStatus.NECROTIC,
+                PhagocyteStatus.APOPTOTIC,
+                PhagocyteStatus.DEAD,
+            }:
+                iron.grid[tuple(neutrophil_cell_voxel)] += neutrophil_cell['iron_pool']
+                neutrophil_cell['iron_pool'] = 0
+                neutrophil_cell['dead'] = True
 
-        n_cells.damage_hyphae(
-            neutrophil.n_det, neutrophil.n_kill, neutrophil.time_n, health, grid, fungus, iron
-        )
+            # interact with fungus
+            if not neutrophil_cell['engaged'] and neutrophil_cell['status'] not in {
+                PhagocyteStatus.APOPTOTIC,
+                PhagocyteStatus.NECROTIC,
+                PhagocyteStatus.DEAD,
+            }:
+                # get fungal cells in this voxel
+                local_aspergillus = afumigatus.cells.get_cells_in_voxel(neutrophil_cell_voxel)
+                for aspergillus_index in local_aspergillus:
+                    aspergillus_cell: AfumigatusCellData = afumigatus.cells[aspergillus_index]
+                    if aspergillus_cell['dead']:
+                        continue
 
-        # update granule == 0 status
-        n_cells.update()
+                    if aspergillus_cell['status'] in {
+                        AfumigatusCellStatus.HYPHAE,
+                        AfumigatusCellStatus.GERM_TUBE,
+                    }:
+                        # possibly internalize the fungal cell
+                        if rg.uniform() < neutrophil.pr_n_hyphae:
+                            internalize_aspergillus(
+                                phagocyte_cell=neutrophil_cell,
+                                aspergillus_cell=aspergillus_cell,
+                                aspergillus_cell_index=aspergillus_index,
+                                phagocyte=neutrophil,
+                            )
+                            aspergillus_cell['status'] = AfumigatusCellStatus.DYING
+                        else:
+                            neutrophil_cell['engaged'] = True
 
-        n_cells.age()
+                    elif aspergillus_cell['status'] == AfumigatusCellStatus.SWELLING_CONIDIA:
+                        if rg.uniform() < neutrophil.pr_n_phagocyte:
+                            internalize_aspergillus(
+                                phagocyte_cell=neutrophil_cell,
+                                aspergillus_cell=aspergillus_cell,
+                                aspergillus_cell_index=aspergillus_index,
+                                phagocyte=neutrophil,
+                            )
+                        else:
+                            pass
 
-        n_cells.kill_by_age(neutrophil.age_limit)
+            # interact with macrophages:
+            # if we are apoptotic, give our iron and phagosome to a nearby present macrophage (if empty)
+            if neutrophil_cell['status'] == PhagocyteStatus.APOPTOTIC:
+                local_macrophages = macrophage.cells.get_cells_in_voxel(neutrophil_cell_voxel)
+                for macrophage_index in local_macrophages:
+                    macrophage_cell: MacrophageCellData = macrophage.cells[macrophage_index]
+                    macrophage_num_cells_in_phagosome = np.sum(macrophage_cell['phagosome'] >= 0)
+                    # TODO: Henrique, why only if empty?
+                    if macrophage_num_cells_in_phagosome == 0:
+                        macrophage_cell['phagosome'] = neutrophil_cell['phagosome']
+                        macrophage_cell['iron_pool'] += neutrophil_cell['iron_pool']
+                        neutrophil_cell[
+                            'iron_pool'
+                        ] = 0.0  # TODO: verify, Henrique's code looks odd
+                        neutrophil_cell['status'] = PhagocyteStatus.DEAD
+                        macrophage_cell['status'] = PhagocyteStatus.INACTIVE
+
+            # Movement
+            if neutrophil_cell['status'] == PhagocyteStatus.ACTIVE:
+                max_move_step = neutrophil.n_move_rate_act * self.time_step
+            else:
+                max_move_step = neutrophil.n_move_rate_rest * self.time_step
+            move_step: int = rg.poisson(max_move_step)  # TODO: verify
+            for _ in range(move_step):
+                self.single_step_move(state, neutrophil_cell)
+
+        # Recruitment
+        self.recruit_neutrophils(state, space_volume, voxel_volume)
 
         return state
 
@@ -302,8 +257,188 @@ class Neutrophil(ModuleModel):
 
         return {
             'count': len(neutrophil.cells.alive()),
-            'granules': int(neutrophil.granule_count),
         }
 
     def visualization_data(self, state: State) -> Tuple[str, Any]:
         return 'cells', state.neutrophil.cells
+
+    def single_step_probabilistic_drift(
+        self, state: State, cell: NeutrophilCellData, voxel: Voxel
+    ) -> Voxel:
+        """
+        Calculate a 1-step voxel movement of a neutrophil
+
+        Parameters
+        ----------
+        state : State
+            global simulation state
+        cell : NeutrophilCellData
+            a neutrophil cell
+        voxel : Voxel
+            current voxel position of the neutrophil
+
+        Returns
+        -------
+        Voxel
+            the new voxel position of the neutrophil
+        """
+        # neutrophils are attracted by MIP2
+
+        neutrophil: NeutrophilState = state.neutrophil
+        mip2: MIP2State = state.mip1b
+        grid: RectangularGrid = state.grid
+        lung_tissue: np.ndarray = state.lung_tissue
+        voxel_volume: float = state.voxel_volume
+
+        # neutrophil has a non-zero probability of moving into non-air voxels
+        nearby_voxels: Tuple[Voxel] = tuple(grid.get_adjacent_voxels(voxel))
+        weights = np.array(
+            [
+                activation_function(
+                    x=mip2.grid[tuple(vxl)], kd=mip2.k_d, h=self.time_step / 60, volume=voxel_volume
+                )
+                + neutrophil.drift_bias
+                if lung_tissue[tuple(vxl)] != TissueType.AIR
+                else 0.0
+                for vxl in nearby_voxels
+            ],
+            dtype=np.float64,
+        )
+
+        normalized_weights = weights / np.sum(weights)
+
+        # sample from distribution given by normalized weights
+        r = rg.uniform()
+        for vxl, weight in zip(nearby_voxels, normalized_weights):
+            if r <= weight:
+                return vxl
+            r -= weight
+
+        assert False, "Sum of normalized weights must be ==1.0, but somehow it isn't."
+
+    def update_status(self, state: State, neutrophil_cell: NeutrophilCellData) -> None:
+        """
+        Update the status of the cell, progressing between states after a certain number of ticks.
+
+        Parameters
+        ----------
+        state : State
+            global simulation state
+        neutrophil_cell : NeutrophilCellData
+
+        Returns
+        -------
+        nothing
+        """
+        neutrophil: NeutrophilState = state.neutrophil
+
+        if neutrophil_cell['status'] in {PhagocyteStatus.NECROTIC, PhagocyteStatus.APOPTOTIC}:
+            self.release_phagosome(state, neutrophil_cell)
+
+        elif rg.uniform() < neutrophil.half_life:
+            neutrophil_cell['status'] = PhagocyteStatus.APOPTOTIC
+
+        elif neutrophil_cell['status'] == PhagocyteStatus.ACTIVE:
+            if neutrophil_cell['status_iteration'] >= neutrophil.iter_to_change_state:
+                neutrophil_cell['status_iteration'] = 0
+                neutrophil_cell['tnfa'] = False
+                neutrophil_cell['status'] = PhagocyteStatus.RESTING
+                neutrophil_cell['state'] = PhagocyteState.FREE  # TODO: was not part of macrophage
+            else:
+                neutrophil_cell['status_iteration'] += 1
+
+        elif neutrophil_cell['status'] == PhagocyteStatus.ACTIVATING:
+            if neutrophil_cell['status_iteration'] >= neutrophil.iter_to_change_state:
+                neutrophil_cell['status_iteration'] = 0
+                neutrophil_cell['status'] = PhagocyteStatus.ACTIVE
+            else:
+                neutrophil_cell['status_iteration'] += 1
+
+    def recruit_neutrophils(self, state: State, space_volume: float, voxel_volume: float) -> None:
+        """
+        Recruit neutrophils based on MIP2 activation
+
+        Parameters
+        ----------
+        state : State
+            global simulation state
+        space_volume : float
+        voxel_volume : float
+
+        Returns
+        -------
+        nothing
+        """
+        from nlisim.modules.mip2 import MIP2State
+
+        neutrophil: NeutrophilState = state.neutrophil
+        mip2: MIP2State = state.mip2
+
+        # 1. compute number of neutrophils to recruit
+        num_live_neutrophils = len(neutrophil.cells.alive())
+        avg = (
+            neutrophil.recruitment_rate
+            * neutrophil.n_frac
+            * np.sum(mip2.grid)
+            * (1 - num_live_neutrophils / neutrophil.max_n)
+            / (mip2.k_d * space_volume)
+        )
+        number_to_recruit = np.random.poisson(avg) if avg > 0 else 0
+        # 2. get voxels for new macrophages, based on activation
+        if number_to_recruit > 0:
+            activation_voxels = zip(
+                *np.where(
+                    rg.uniform(size=mip2.grid.shape)
+                    < activation_function(
+                        x=mip2.grid,
+                        kd=mip2.k_d,
+                        h=self.time_step / 60,
+                        volume=voxel_volume,
+                        b=neutrophil.rec_bias,
+                    )
+                )
+            )
+            for coordinates in rg.choice(activation_voxels, size=number_to_recruit, replace=True):
+                z, y, x = coordinates + rg.uniform(3)  # TODO: discuss placement
+                self.create_neutrophil(state, x, y, z)
+                # TODO: have placement fail due to overcrowding of cells
+
+    @staticmethod
+    def create_neutrophil(state: State, x: float, y: float, z: float, **kwargs) -> None:
+        """
+        Create a new neutrophil cell
+
+        Parameters
+        ----------
+        state : State
+            global simulation state
+        x : float
+        y : float
+        z : float
+            coordinates of created neutrophil
+        kwargs
+            parameters for neutrophil, will give
+
+        Returns
+        -------
+        nothing
+        """
+        neutrophil: NeutrophilState = state.neutrophil
+
+        if 'iron_pool' in kwargs:
+            neutrophil.cells.append(
+                NeutrophilCellData.create_cell(point=Point(x=x, y=y, z=z), **kwargs)
+            )
+        else:
+            neutrophil.cells.append(
+                NeutrophilCellData.create_cell(point=Point(x=x, y=y, z=z), iron_pool=0.0, **kwargs)
+            )
+
+
+# def get_max_move_steps(self):  ##REVIEW
+#     if self.max_move_step is None:
+#         if self.status == Macrophage.ACTIVE:
+#             self.max_move_step = np.random.poisson(Constants.MA_MOVE_RATE_REST)
+#         else:
+#             self.max_move_step = np.random.poisson(Constants.MA_MOVE_RATE_REST)
+#     return self.max_move_step
