+++ nlisim/modules/tafc.py
@@ -5,7 +5,7 @@ import numpy as np
 
 from nlisim.coordinates import Voxel
 from nlisim.grid import RectangularGrid
-from nlisim.henrique_modules.molecules import MoleculeModel, MoleculesState
+from nlisim.modules.molecules import MoleculeModel, MoleculesState
 from nlisim.module import ModuleState
 from nlisim.state import State
 from nlisim.util import michaelian_kinetics, turnover_rate
@@ -50,9 +50,9 @@ class TAFC(MoleculeModel):
 
     def advance(self, state: State, previous_time: float) -> State:
         """Advance the state by a single time step."""
-        from nlisim.henrique_modules.iron import IronState
-        from nlisim.henrique_modules.transferrin import TransferrinState
-        from nlisim.henrique_modules.afumigatus import (
+        from nlisim.modules.iron import IronState
+        from nlisim.modules.transferrin import TransferrinState
+        from nlisim.modules.afumigatus import (
             AfumigatusCellData,
             AfumigatusCellState,
             AfumigatusCellStatus,
