+++ nlisim/modules/tnfa.py
@@ -6,7 +6,7 @@ import numpy as np
 
 from nlisim.coordinates import Voxel
 from nlisim.grid import RectangularGrid
-from nlisim.henrique_modules.molecules import MoleculeModel, MoleculesState
+from nlisim.modules.molecules import MoleculeModel, MoleculesState
 from nlisim.module import ModuleState
 from nlisim.random import rg
 from nlisim.state import State
@@ -56,9 +56,9 @@ class TNFa(MoleculeModel):
 
     def advance(self, state: State, previous_time: float) -> State:
         """Advance the state by a single time step."""
-        from nlisim.henrique_modules.macrophage import MacrophageCellData, MacrophageState
-        from nlisim.henrique_modules.neutrophil import NeutrophilCellData, NeutrophilState
-        from nlisim.henrique_modules.phagocyte import PhagocyteStatus
+        from nlisim.modules.macrophage import MacrophageCellData, MacrophageState
+        from nlisim.modules.neutrophil import NeutrophilCellData, NeutrophilState
+        from nlisim.modules.phagocyte import PhagocyteStatus
 
         tnfa: TNFaState = state.antitnfa
         molecules: MoleculesState = state.molecules
