+++ nlisim/oldmodules/afumigatus.py-deprecated
