+++ nlisim/oldmodules/state_output.py
@@ -1,4 +1,3 @@
-from typing import Any, Dict, Tuple
 from pathlib import Path
 import shutil
