+++ test/test_epithelium.py
@@ -3,8 +3,8 @@ from pytest import fixture
 
 from nlisim.coordinates import Point
 from nlisim.grid import RectangularGrid
-from nlisim.modules.epithelium import EpitheliumCellData, EpitheliumCellList
-from nlisim.modules.fungus import FungusCellData, FungusCellList
+from nlisim.oldmodules.epithelium import EpitheliumCellData, EpitheliumCellList
+from nlisim.oldmodules.fungus import FungusCellData, FungusCellList
 
 
 @fixture
