+++ test/test_fungus.py
@@ -3,7 +3,7 @@ from pytest import fixture
 
 from nlisim.coordinates import Point
 from nlisim.grid import RectangularGrid
-from nlisim.modules.fungus import FungusCellData, FungusCellList
+from nlisim.oldmodules.fungus import FungusCellData, FungusCellList
 
 
 @fixture
